move some registration requirements out of the component to avoid nested ternaries
